Add files via upload
It was a great learning experience to analyze the 'International debt data' collected by The World Bank. The dataset contains information about the amount of debt (in USD) owed by developing countries across several categories. Performed Data extraction and cleaning along with Data manipulation.

 There were six indicators name and knowing about these various debt indicators helped understand the areas in which a country can possibly be indebted to. Also, the queries helped find out how much debt a country owes on an average as well as gave a better sense of the distribution of the amount of debt across different indicators. I could see that the indicator DT.AMT.DLXF.CD topped the chart of average debt. This category includes 'repayment of long term debts'. Countries take on long-term debt to acquire immediate capital.

Since not all the countries suffer from the same kind of economic disturbances, this finding allowed me to understand a particular country's economic condition a bit more specifically. Hence, this gave me a clue that all these countries are suffering from a common economic issue. I saw that long-term debt is the topmost category when it comes to the average amount of debt.